Add methods for inserting POIs to repo class

+++ app/app/src/main/java/net/easimer/surveyor/data/disk/entities/PointOfInterest.kt
@@ -5,8 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class PointOfInterest(
-    @PrimaryKey(autoGenerate = true) val poiId: Int,
-    val recId: Int,
+    @PrimaryKey(autoGenerate = true) val poiId: Long,
+    val recId: Long,
     val title: String,
     val longitude: Double,
     val latitude: Double
